+++ src/main/java/org/sciborgs1155/robot/commands/GroundIntake.java
@@ -3,14 +3,14 @@ package org.sciborgs1155.robot.commands;
 import edu.wpi.first.wpilibj2.command.Command;
 import org.sciborgs1155.robot.arm.Arm;
 import org.sciborgs1155.robot.arm.ArmConstants;
-import org.sciborgs1155.robot.roller.Roller;
+import org.sciborgs1155.robot.coroller.Coroller;
 
 public class GroundIntake {
 
   private Arm arm;
-  private Roller roller;
+  private Coroller roller;
 
-  public GroundIntake(Arm arm, Roller roller) {
+  public GroundIntake(Arm arm, Coroller roller) {
     this.arm = arm;
     this.roller = roller;
   }
