+++ src/main/java/org/sciborgs1155/robot/coroller/CorollerConstants.java
@@ -1,7 +1,7 @@
-package org.sciborgs1155.robot.roller;
+package org.sciborgs1155.robot.coroller;
 
-/** Constants for the {@link Roller} subsystem. */
-public class RollerConstants {
+/** Constants for the {@link Coroller} subsystem. */
+public class CorollerConstants {
   // The amount of power going into the roller (from -1 to 1)
   public static final double INTAKE_POWER = 0.5; // TODO: arbitrary value
   public static final double OUTTAKE_POWER = -0.2; // TODO: also arbitrary
